CastObject module
